address comments
still not sure how to handle ambiguous case where the user provides only
the group name but intends to load an existing model. In this case I
think what currently will happen is that a new directory/model will be
made? we need to have code that checks to make sure a checkpoint exists
before doing evaluation or something